requirements-offline-linux.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
configparser==5.3.0
cx-Oracle==8.3.0
exceptiongroup==1.1.0
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.23.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
patsy==0.5.3
pluggy==1.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.2.0
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scikit-surprise==1.1.3
scipy==1.10.0
six==1.16.0
statsmodels==0.13.5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
